+++ tests/team/test_client.py
@@ -77,3 +77,18 @@ def test_update_team_attribute(requests_mock):
         status_code=200,
     )
     assert TeamClient().update_team_attribute(team_id, "a", "b") == team
+
+
+def test_update_team_members(requests_mock):
+    team_id = "abc"
+    members = [
+        {"name": "Foo", "username": "foo"},
+        {"name": "Bar", "username": "bar"},
+    ]
+    requests_mock.register_uri(
+        "PUT",
+        re.compile(f"teams/{team_id}/members"),
+        text=json.dumps(members),
+        status_code=200,
+    )
+    assert TeamClient().update_team_members(team_id, ["abc-123", "def-456"]) == members