+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## ?.?.?
+
+* Add method `TeamClient.update_team_members` for adding and/or removing
+  members to/from a team.
+
 ## 2.2.1
 
 * Fixed a bug in `TeamClient.update_team_attribute` when `value` is falsy.