+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## ?.?.?
+## 2.3.0
 
 * New method `TeamClient.get_user_by_username` for looking up Keycloak users.
 * New method `TeamClient.update_team_members` for adding and/or removing